+++ server/routes/router.js
@@ -1,26 +1,27 @@
 import express from "express";
 
-import { PostController, UserController } from "../controllers/Controllers";
+import { PostController } from "../controllers/PostControllers";
+import { UserController } from "../controllers/UserControllers";
 
 import verify from "../middlewares/verifyToken";
 const router = express.Router();
 
 // Retrieve a list of all articles
-router.get("/posts", PostController.retriveArticles);
+router.get("/articles", PostController.retriveArticles);
 
 router.post("/signup", UserController.createUser);
 
 router.post("/login", UserController.loginUser);
 // create an article
-router.post("/posts/create", verify, PostController.createPost);
+router.post("/articles/create", verify, PostController.createPost);
 // Retrieve a single article
-router.get("/posts/:id", verify, PostController.retrieveOnePost);
+router.get("/articles/:id", verify, PostController.retrieveOnePost);
 // create comments
-router.post("/posts/:id/comments", PostController.comments);
+router.post("/articles/:id/comments", PostController.comments);
 
 // Update an existing article
-router.patch("/posts/:id", verify, PostController.updateArticle);
+router.patch("/articles/:id", verify, PostController.updateArticle);
 // Delete an existing article
-router.delete("/posts/:id", verify, PostController.deleteArticle);
+router.delete("/articles/:id", verify, PostController.deleteArticle);
 
 export default router;
